+++ tests/apiv1/test_sample_groups.py
@@ -3,6 +3,8 @@
 import json
 from uuid import UUID
 
+from sqlalchemy.orm.exc import NoResultFound
+
 from app import db
 from app.display_modules.ancestry.constants import TOOL_MODULE_NAME
 from app.samples.sample_models import Sample
@@ -77,6 +79,62 @@ class TestSampleGroupModule(BaseTestCase):
         organization = add_organization('Organization', 'admin@organization.org')
         response = self.create_group_for_organization(auth_headers, organization.id)
         self.assertEqual(response.status_code, 403)
+
+    def delete_sample_group(self, auth_headers, sample_group_id):
+        """Perform request to delete sample group."""
+        with self.client:
+            response = self.client.delete(
+                f'/api/v1/sample_groups/{sample_group_id}',
+                headers=auth_headers,
+                content_type='application/json',
+            )
+            return response
+
+    @with_user
+    def test_delete_sample_group(self, auth_headers, *_):
+        """Ensure an unowned sample group can be removed from the database."""
+        sample_group = add_sample_group(name='The Least Sampled of Groups')
+        response = self.delete_sample_group(auth_headers, sample_group.id)
+        data = json.loads(response.data.decode())
+        self.assertEqual(response.status_code, 200)
+        self.assertIn('success', data['status'])
+
+        # Ensure Sample Group was removed
+        query = SampleGroup.query.filter_by(id=sample_group.id)
+        self.assertRaises(NoResultFound, query.one)
+
+    @with_user
+    def test_delete_owned_sample_group(self, auth_headers, login_user):
+        """Ensure an owned sample group can be removed from the database by an authorized user."""
+        organization = add_organization('Organization', 'admin@organization.org')
+        organization.users.append(login_user)
+        sample_group = add_sample_group(name='Owned Sample Group')
+        organization.sample_groups.append(sample_group)
+        db.session.commit()
+        response = self.delete_sample_group(auth_headers, sample_group.id)
+        data = json.loads(response.data.decode())
+        self.assertEqual(response.status_code, 200)
+        self.assertIn('success', data['status'])
+
+        # Ensure Sample Group was removed
+        query = SampleGroup.query.filter_by(id=sample_group.id)
+        self.assertRaises(NoResultFound, query.one)
+
+    @with_user
+    def test_unauthorized_delete_sample_group(self, auth_headers, *_):  # pylint: disable=invalid-name
+        """Ensure an owned sample group cannot be removed by an unauthorized user."""
+        organization = add_organization('Organization', 'admin@organization.org')
+        sample_group = add_sample_group(name='Owned Sample Group')
+        organization.sample_groups.append(sample_group)
+        db.session.commit()
+        response = self.delete_sample_group(auth_headers, sample_group.id)
+        data = json.loads(response.data.decode())
+        self.assertEqual(response.status_code, 403)
+        self.assertIn('error', data['status'])
+
+        # Ensure Sample Group still exists
+        group = SampleGroup.query.filter_by(id=sample_group.id).one()
+        self.assertTrue(group)
 
     @with_user
     def test_add_samples_to_group(self, auth_headers, *_):